test(ci-speed): split tank_crit and professions_trend for N=8 balance
Phase 3 batch 1: pure describe/class-pair moves so the two worst N=8
monsters leave shards 4 and 5. tank_crit_immunity (CI ~174s on s4)
becomes three class-pair files plus util; professions_trend (CI ~105s
on s5) becomes five describe files plus util. Assertion bodies and
test count (32) unchanged. Simulated sha1 destinations: warrior/paladin/
druid pairs on s2/s1/s6; trend pieces on s1/s7/s5/s6/s1 (none stacked
back on both worst shards).
