+++ tests/tank_crit_immunity_util.ts
@@ -1,4 +1,9 @@
-import { describe, expect, it, vi } from 'vitest';
+// Shared fight helper for the tank crit-immunity suite split
+// (tank_crit_immunity_*_pair.test.ts). Not a test file: nothing here runs
+// on its own. Split from the original single-file tests/tank_crit_immunity.test.ts
+// along class-pair boundaries for CI shard balance (a pure move).
+
+import { expect, vi } from 'vitest';
 import { MOBS } from '../src/sim/data';
 import { createMob } from '../src/sim/entity';
 import { Sim } from '../src/sim/sim';
@@ -18,10 +23,10 @@ vi.setConfig({ testTimeout: 120000 });
 // else keeps eating 2x mob crits, and the 5% crit roll is still DRAWN for the
 // immune tank so every downstream rng draw keeps its stream position.
 
-const SEED = 90210;
-const WINDOW_SECONDS = 240;
+export const SEED = 90210;
+export const WINDOW_SECONDS = 240;
 
-type Setup = {
+export type Setup = {
   cls: PlayerClass;
   spec: string | null;
   form?: 'bear_form';
@@ -29,7 +34,7 @@ type Setup = {
 
 // One identical fight per case: same seed, same mob, same window; only the
 // defender's build differs. Returns landed swings and crits taken.
-function critsTaken(setup: Setup): { hits: number; crits: number } {
+export function critsTaken(setup: Setup): { hits: number; crits: number } {
   const sim = new Sim({ seed: SEED, playerClass: 'warrior', noPlayer: true });
   const pid = sim.addPlayer(setup.cls, 'Defender');
   sim.setPlayerLevel(20, pid);
@@ -78,29 +83,3 @@ function critsTaken(setup: Setup): { hits: number; crits: number } {
   expect(hits).toBeGreaterThan(30); // the fight actually ran
   return { hits, crits };
 }
-
-describe('tank crit immunity vs mobs', () => {
-  it('a Protection warrior is never critically hit', () => {
-    expect(critsTaken({ cls: 'warrior', spec: 'prot' }).crits).toBe(0);
-  });
-
-  it('an Arms warrior still eats mob crits (the roll is alive)', () => {
-    expect(critsTaken({ cls: 'warrior', spec: 'arms' }).crits).toBeGreaterThan(0);
-  });
-
-  it('a Protection paladin is never critically hit', () => {
-    expect(critsTaken({ cls: 'paladin', spec: 'protection' }).crits).toBe(0);
-  });
-
-  it('a Retribution paladin still eats mob crits', () => {
-    expect(critsTaken({ cls: 'paladin', spec: 'retribution' }).crits).toBeGreaterThan(0);
-  });
-
-  it('a Feral druid in Sloth Form is never critically hit', () => {
-    expect(critsTaken({ cls: 'druid', spec: 'feral', form: 'bear_form' }).crits).toBe(0);
-  });
-
-  it('a Feral druid OUT of form still eats mob crits: the form is the commitment', () => {
-    expect(critsTaken({ cls: 'druid', spec: 'feral' }).crits).toBeGreaterThan(0);
-  });
-});
